services/db: inline UserSettings model
